requirements.txt
flask
flask-cors
ultralytics
opencv-python
torch
torchvision
numpy
Pillow
gunicorn
